% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{AtRiskMatrixR}
\alias{AtRiskMatrixR}
\title{Tabulate At-Risk Matrix R}
\usage{
AtRiskMatrixR(eval_times, idx, time)
}
\arguments{
\item{eval_times}{Evaluation times.}

\item{idx}{Unique subject index.}

\item{time}{Observation time.}
}
\value{
Numeric matrix.
}
\description{
Construct a subject (row) by evaluation time (col) matrix.
}
