% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\docType{class}
\name{AURMC-class}
\alias{AURMC-class}
\title{AURMC Object}
\description{
AURMC Object
}
\section{Slots}{

\describe{
\item{\code{Arm0}}{Results for the reference arm.}

\item{\code{Arm1}}{Results for the treatment arm.}

\item{\code{Contrast}}{Contrasts.}
}}

