% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{show,AURMC-method}
\alias{show,AURMC-method}
\title{Show Method for AURMC Object}
\usage{
\S4method{show}{AURMC}(object)
}
\arguments{
\item{object}{An object of class \code{AURMC}.}
}
\description{
Show Method for AURMC Object
}
