% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneSample.R
\name{AURMC}
\alias{AURMC}
\title{Area Under the Repeated Measures Curve}
\usage{
AURMC(
  data,
  alpha = 0.05,
  censor_after_last = TRUE,
  idx_name = "idx",
  int_method = "trapezoid",
  perturbations = NULL,
  random_state = 0,
  status_name = "status",
  tau = NULL,
  time_name = "time",
  value_name = "value"
)
}
\arguments{
\item{data}{Data.frame.}

\item{alpha}{Type I error.}

\item{censor_after_last}{Introduce censoring after the last event \emph{if} no
observation-terminating event is present.}

\item{idx_name}{Name of column containing a unique subject index.}

\item{int_method}{Integration method, selected from "left", "right", "trapezoid".}

\item{perturbations}{Number of perturbations to use for bootstrap inference.
If \code{NULL}, only analytical inference is performed.}

\item{random_state}{Seed to ensure perturbations are reproducible.}

\item{status_name}{Name of column containing the status. Must be coded as 0
for censoring, 1 for a measurement, 2 for death. Each subject should have
an observation-terminating event, either censoring or death.}

\item{tau}{Truncation time.}

\item{time_name}{Name of column containing the observation time.}

\item{value_name}{Name of the column containing the measurement.}
}
\value{
A data.frame containing the method, truncation time, area estimate,
standard error, confidence limits, and p-value.
}
\description{
Area Under the Repeated Measures Curve
}
\examples{
example_data <- data.frame(
  idx = rep(1:3, each = 2),
  time = rep(c(0, 1), 3),
  status = rep(c(1, 0), 3),
  value = c(1, 1, 2, 2, 3, 3)
)
AURMC(example_data, tau = 1)
}
