% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CalcMartingaleR}
\alias{CalcMartingaleR}
\title{Calculate Martingale}
\usage{
CalcMartingaleR(haz, idx, status, time, unique_times)
}
\arguments{
\item{haz}{Value of the hazard at each unique time.}

\item{idx}{Subject index.}

\item{status}{Status, coded as 0 for censoring, 1 for a measurement, 2 for a terminal event.}

\item{time}{Subject observation times.}

\item{unique_times}{Unique times at which to obtain the martingale.}
}
\value{
Matrix with subjects as rows and unique times as columns.
}
\description{
Calculate \eqn{dM_{i}(t) = dN_{i}(t) - Y_{i}(t)d\Lambda(t)}.
}
