% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputCheck.R
\name{CensorAfterLast}
\alias{CensorAfterLast}
\title{Censor After Last}
\usage{
CensorAfterLast(data)
}
\arguments{
\item{data}{Data.frame.}
}
\value{
None.
}
\description{
Introduce a censoring after the last event if no observation
terminating event is present.
}
