% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Curve.R
\name{Curve}
\alias{Curve}
\title{Estimate Curve}
\usage{
Curve(
  data,
  censor_after_last = TRUE,
  idx_name = "idx",
  status_name = "status",
  time_name = "time",
  value_name = "value"
)
}
\arguments{
\item{data}{Data.frame.}

\item{censor_after_last}{Introduce censoring after the last event \emph{if} no
observation-terminating event is present.}

\item{idx_name}{Name of column containing a unique subject index.}

\item{status_name}{Name of column containing the status. Must be coded as 0
for censoring, 1 for a measurement, 2 for death. Each subject should have
an observation-terminating event, either censoring or death.}

\item{time_name}{Name of column containing the observation time.}

\item{value_name}{Name of the column containing the measurement.}
}
\value{
Function.
}
\description{
Estimate Curve
}
