% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{InterpolateR}
\alias{InterpolateR}
\title{Interpolation R

Linearly interpolates between each subject's measurements.
The input data should contain no missing values.}
\usage{
InterpolateR(grid, idx, status, time, value)
}
\arguments{
\item{grid}{Grid of unique points at which to interpolate.}

\item{idx}{Unique subject index.}

\item{status}{Status, coded as 0 for censoring, 1 for a measurement, 2 for a terminal event.}

\item{time}{Observation time.}

\item{value}{Observation value.}
}
\value{
Data.frame.
}
\description{
Interpolation R

Linearly interpolates between each subject's measurements.
The input data should contain no missing values.
}
