% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{PlotTwoSample}
\alias{PlotTwoSample}
\title{Plot Two Sample}
\usage{
PlotTwoSample(
  data,
  arm_name = "arm",
  censor_after_last = TRUE,
  color_labs = c("Ctrl", "Trt"),
  colors = c("#EFC000FF", "#6385B8"),
  eval_pts = 200,
  idx_name = "idx",
  show_auc = FALSE,
  show_nar = TRUE,
  status_name = "status",
  tau = NULL,
  time_name = "time",
  title = NULL,
  value_name = "value",
  x_breaks = NULL,
  x_labs = NULL,
  x_name = "Time",
  x_max = NULL,
  y_name = "Expected Value",
  y_lim = NULL
)
}
\arguments{
\item{data}{Data.frame.}

\item{arm_name}{Name of the column containing treatment arm. Coded as 0 for
reference, 1 for treatment.}

\item{censor_after_last}{Introduce censoring after the last event \emph{if} no
observation-terminating event is present.}

\item{color_labs}{Color labels.}

\item{colors}{Colors for control and treatment respectively.}

\item{eval_pts}{Evaluation points between 0 and tau.}

\item{show_auc}{Show area under the curve?}

\item{show_nar}{Show number at risk?}

\item{status_name}{Name of column containing the status. Must be coded as 0
for censoring, 1 for a measurement, 2 for death. Each subject should have
an observation-terminating event, either censoring or death.}

\item{tau}{Truncation time.}

\item{time_name}{Name of column containing the observation time.}

\item{title}{Plot title.}

\item{value_name}{Name of the column containing the measurement.}

\item{x_breaks}{X-axis breaks.}

\item{x_labs}{X-axis labels.}

\item{x_name}{X-axis name.}

\item{x_max}{X-axis upper limit.}

\item{y_name}{Y-axis name.}

\item{y_lim}{Y-axis limits.}
}
\value{
ggplot
}
\description{
Plot Two Sample
}
