% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{print.AURMC}
\alias{print.AURMC}
\title{Print Method for AURMC Object}
\usage{
\method{print}{AURMC}(x, ...)
}
\arguments{
\item{x}{An object of class \code{AURMC}.}

\item{...}{Unused.}
}
\description{
Print method for objects of class \code{AURMC}.
}
