% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ValueMatrixR}
\alias{ValueMatrixR}
\title{Tabulate Value Matrix R}
\usage{
ValueMatrixR(eval_times, idx, time, value)
}
\arguments{
\item{eval_times}{Evaluation times.}

\item{idx}{Unique subject index.}

\item{time}{Observation time.}

\item{value}{Observation value.}
}
\value{
Numeric matrix.
}
\description{
Tabulate \eqn{D_{i}(t)} with subjects as rows and evaluation times as columns.
}
