% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AURMC-package.R
\docType{package}
\name{AURMC-package}
\alias{AURMC-package}
\alias{_PACKAGE}
\title{AURMC: Area Under the Repeated Measures Curve}
\description{
Inference on the area under the repeated measures curve.
}
\author{
\strong{Maintainer}: Zachary McCaw \email{zmccaw@alumni.harvard.edu} (\href{https://orcid.org/0000-0002-2006-9828}{ORCID})

}
\keyword{internal}
