% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BootstrapSamplesR}
\alias{BootstrapSamplesR}
\title{Bootstrap Samples R}
\usage{
BootstrapSamplesR(
  boot,
  eval_times,
  idx,
  status,
  time,
  value,
  return_auc = FALSE
)
}
\arguments{
\item{boot}{Bootstrap replicates.}

\item{eval_times}{Evaluation times.}

\item{idx}{Unique subject index.}

\item{status}{Status, coded as 0 for censoring, 1 for event.}

\item{time}{Observation time.}

\item{value}{Observation value.}

\item{return_auc}{Return the AUC?}
}
\value{
Numeric matrix.
}
\description{
Bootstrap Samples R
}
