% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BootstrapSamplesR}
\alias{BootstrapSamplesR}
\title{Bootstrap Samples R}
\usage{
BootstrapSamplesR(
  boot,
  eval_times,
  idx,
  status,
  time,
  value,
  replace_na = FALSE,
  return_auc = FALSE,
  trunc_time = NULL
)
}
\arguments{
\item{boot}{Bootstrap replicates.}

\item{eval_times}{Evaluation times.}

\item{idx}{Unique subject index.}

\item{status}{Status, coded as 0 for censoring, 1 for event, 2 for terminal event.}

\item{time}{Observation time.}

\item{value}{Observation value.}

\item{replace_na}{Replace NaN with zero?}

\item{return_auc}{Return the AUC?}

\item{trunc_time}{Truncation time? Optional. If omitted, defaults
to the maximum evaluation time.}
}
\value{
Numeric matrix.
}
\description{
Returns a bootstrap (row) by statistic (col) matrix. If \code{return_auc = TRUE},
the output has a single column, the AUC. If \code{return_auc = FALSE}, the output
has a single column for each evaluation time.
}
