% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CalcMuR}
\alias{CalcMuR}
\title{Calculate Mu R}
\usage{
CalcMuR(d, surv, unique_times, y)
}
\arguments{
\item{d}{Value of d(t) at each time point.}

\item{surv}{Value of S(t) at each time point.}

\item{unique_times}{Unique values of time t.}

\item{y}{Value of y(t) at each time point.}
}
\value{
Numeric vector of \eqn{\mu(t; tau)}.
}
\description{
Evaluate \eqn{\mu(t; \tau) = \int_{t}^{\tau}{S(u)d(u)/y(u)}du}.
}
