% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DrawBootstrapR}
\alias{DrawBootstrapR}
\title{Draw Bootstrap R}
\usage{
DrawBootstrapR(idx, status, time, value)
}
\arguments{
\item{idx}{Unique subject index.}

\item{status}{Status, coded as 0 for censoring, 1 for event.}

\item{time}{Observation time.}

\item{value}{Observation value.}
}
\value{
Numeric matrix.
}
\description{
Draw Bootstrap R
}
