% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EstimatorR}
\alias{EstimatorR}
\title{Tabulate Estimator R}
\usage{
EstimatorR(
  idx,
  status,
  time,
  value,
  eval_times = NULL,
  int_method = "trapezoid",
  replace_na = FALSE,
  return_auc = FALSE,
  trunc_time = NULL
)
}
\arguments{
\item{idx}{Unique subject index.}

\item{status}{Status, coded as 0 for censoring, 1 for event, 2 for terminal event.}

\item{time}{Observation time.}

\item{value}{Observation value.}

\item{eval_times}{Evaluation times. If omitted, defaults to the
unique values of time.}

\item{int_method}{Integration method, selected from "left", "right", "trapezoid".}

\item{replace_na}{Replace NaN with zero? Default: FALSE.}

\item{return_auc}{Return the AUC? Default: FALSE.}

\item{trunc_time}{Truncation time? Optional. If omitted, defaults
to the maximum evaluation time.}
}
\value{
Data.frame.
}
\description{
Tabulate Estimator R
}
