% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EstimatorR}
\alias{EstimatorR}
\title{Tabulate Estimator R}
\usage{
EstimatorR(
  idx,
  status,
  time,
  value,
  eval_times = NULL,
  replace_na = FALSE,
  return_auc = FALSE
)
}
\arguments{
\item{idx}{Unique subject index.}

\item{status}{Status, coded as 0 for censoring, 1 for event.}

\item{time}{Observation time.}

\item{value}{Observation value.}

\item{eval_times}{Times at which to evaluate the estimator.}

\item{replace_na}{Replace NaN with zero?}

\item{return_auc}{Return the AUC?}
}
\value{
Data.frame.
}
\description{
Tabulate Estimator R
}
