% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{GenData}
\alias{GenData}
\title{Simulate Repeated Measures Data}
\usage{
GenData(censoring_rate, death_rate, n, tau)
}
\arguments{
\item{censoring_rate}{Rate for the time to death.}

\item{death_rate}{Rate for the time to death.}

\item{n}{Number of subjects.}

\item{tau}{Truncation time.}
}
\value{
Data.frame.
}
\description{
Simulate Repeated Measures Data
}
