% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{GenData}
\alias{GenData}
\title{Simulate Repeated Measures Data}
\usage{
GenData(
  censoring_rate,
  death_rate,
  n,
  tau,
  last_missing = FALSE,
  value_mean = 0,
  value_sd = 1
)
}
\arguments{
\item{censoring_rate}{Rate for the time to death.}

\item{death_rate}{Rate for the time to death.}

\item{n}{Number of subjects.}

\item{tau}{Truncation time.}

\item{last_missing}{Should the last value be missing? Default: FALSE.}

\item{value_mean}{Mean of measurement.}

\item{value_sd}{Standard deviation of measurement.}
}
\value{
Data.frame.
}
\description{
Simulates repeated measures data at regular time points subject to
exponential censoring and death.
}
