% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{InfluenceR}
\alias{InfluenceR}
\title{Influence Function R}
\usage{
InfluenceR(idx, status, time, trunc_time, value)
}
\arguments{
\item{idx}{Unique subject index.}

\item{status}{Status, coded as 0 for censoring, 1 for event.}

\item{time}{Observation time.}

\item{trunc_time}{Truncation time? Optional. If omitted, defaults
to the maximum evaluation time.}

\item{value}{Observation value.}
}
\value{
Data.frame.
}
\description{
Influence function contributions for the AUC.
}
