% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{InfluenceR}
\alias{InfluenceR}
\title{Influence Function R}
\usage{
InfluenceR(idx, status, time, trunc_time, value)
}
\arguments{
\item{idx}{Unique subject index.}

\item{status}{Status, coded as 0 for censoring, 1 for event, 2 for terminal event.}

\item{time}{Observation time.}

\item{trunc_time}{Truncation time? Optional. If omitted, defaults
to the maximum evaluation time.}

\item{value}{Observation value.}
}
\value{
Data.frame.
}
\description{
Influence function contributions for the AUC. Includes the three component
integrals (\code{i1}, \code{i2}, \code{i3}) and the overall influence \code{psi}.
}
