% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interpolate.R
\name{Interpolate}
\alias{Interpolate}
\title{Interpolate}
\usage{
Interpolate(
  data,
  grid,
  idx_name = "idx",
  status_name = "status",
  rm_na = TRUE,
  time_name = "time",
  value_name = "value"
)
}
\arguments{
\item{data}{Data.frame.}

\item{grid}{Grid of unique points at which to interpolate.}

\item{idx_name}{Name of column containing a unique subject index.}

\item{status_name}{Name of column containing the status. Must be coded as 0
for censoring, 1 for a measurement, 2 for death. Each subject should have
an observation-terminating event, either censoring or death.}

\item{rm_na}{Remove records interpolated to NA?}

\item{time_name}{Name of column containing the observation time.}

\item{value_name}{Name of the column containing the measurement.}
}
\value{
Data.frame.
}
\description{
Linearly interpolations between each subject's measurements.
The input data should contain no missing values.
}
\section{Notes}{

The grid of points will be agumented to include the time of each
subjects observating terminating event.
}

