% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{KaplanMeierR}
\alias{KaplanMeierR}
\title{Tabulate Kaplan Meier R}
\usage{
KaplanMeierR(eval_times, idx, status, time)
}
\arguments{
\item{eval_times}{Evaluation times.}

\item{idx}{Unique subject index.}

\item{status}{Status, 0 for censoring, 1 for event, 2 for death.}

\item{time}{Observation time.}
}
\value{
Data.frame.
}
\description{
Constructs a matrix with evaluation times as rows, and 3 columns, for
time, number at risk, and Kaplan-Meier survival probability.
}
