% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{KaplanMeierR}
\alias{KaplanMeierR}
\title{Tabulate Kaplan Meier R}
\usage{
KaplanMeierR(eval_times, idx, status, time)
}
\arguments{
\item{eval_times}{Evaluation times.}

\item{idx}{Unique subject index.}

\item{status}{Status, coded as 0 for censoring, 1 for event, 2 for terminal event.}

\item{time}{Observation time.}
}
\value{
Data.frame.
}
\description{
Constructs a matrix with evaluation times as rows, and 4 columns:
\itemize{
\item{time}{Evaluation times.}
\item{nar}{Number at risk.}
\item{surv}{Survival probability.}
\item{haz}{Hazard.}
}
}
