% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PerturbationR}
\alias{PerturbationR}
\title{Perturbation R

Generates realizations of \eqn{\frac{1}{n}\sum_{i=1}^{n}\psi_{i}w_{i}},
where \eqn{\psi_{i}} is the influence function for the ith subject and the
\eqn{w_{i}} are IID random weights.}
\usage{
PerturbationR(idx, perturbations, status, time, trunc_time, value)
}
\arguments{
\item{idx}{Unique subject index.}

\item{perturbations}{Number of perturbations}

\item{status}{Status, coded as 0 for censoring, 1 for event.}

\item{time}{Observation time.}

\item{trunc_time}{Truncation time? Optional. If omitted, defaults
to the maximum evaluation time.}

\item{value}{Observation value.}
}
\value{
Numeric vector.
}
\description{
Perturbation R

Generates realizations of \eqn{\frac{1}{n}\sum_{i=1}^{n}\psi_{i}w_{i}},
where \eqn{\psi_{i}} is the influence function for the ith subject and the
\eqn{w_{i}} are IID random weights.
}
\section{Notes}{

The random seed should be set in R prior to calling this function.
}

