% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ValueMatrixR}
\alias{ValueMatrixR}
\title{Tabulate Value Matrix R}
\usage{
ValueMatrixR(eval_times, idx, status, time, value)
}
\arguments{
\item{eval_times}{Evaluation times.}

\item{idx}{Unique subject index.}

\item{status}{Status, 0 for censoring, 1 for event, 2 for death.}

\item{time}{Observation time.}

\item{value}{Observation value.}
}
\value{
Numeric matrix.
}
\description{
Construct a subject (row) by evluation time (col) matrix.
}
